requirements.txt
streamlit==1.31.1
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-mock==3.12.0
apify-client==1.6.1
pydantic==2.6.1
pyperclip==1.8.2